pyproject.toml
[project]
name = "VZOR-cv"
version = "0.0.1"
requires-python = "==3.12.*"
dependencies = [
    "easyocr>=1.7.2",
    "grpcio>=1.72.1",
    "grpcio-tools>=1.72.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "sentencepiece>=0.2.0",
    "seqlog>=0.4.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.52.4",
    "ultralytics==8.3.147",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.12",
]
